Move Loaders, Logos, Navigation components into directories
