+++ assets/javascripts/kitten/components/loaders/loader-with-paragraph/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import classNames from 'classnames'
-import { Loader } from '../../components/loaders/loader'
-import { Paragraph } from '../../components/typography/paragraph'
+import { Loader } from '../../../components/loaders/loader'
+import { Paragraph } from '../../../components/typography/paragraph'
 
 export class LoaderWithParagraph extends Component {
   render() {
