+++ assets/javascripts/kitten/components/loaders/loader/index.js
