+++ assets/javascripts/kitten/components/loaders/loader/test.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Loader } from './loader'
+import { Loader } from './index'
 
 describe('<Loader />', () => {
   describe('by default', () => {
