+++ assets/javascripts/kitten/components/navigation/tab-bar/index.js
